requirements.txt
numpy
scipy
pandas
matplotlib
scikit-learn
tqdm
pyyaml
mne
pyedflib
torch
torchaudio
